:recycle: refactor: migrate application to cli
