+++ src/ServiceBusLifeboat.Domain/ServiceBusLifeboat.Domain.csproj
@@ -6,4 +6,13 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Reference Include="Azure.Core">
+      <HintPath>..\..\..\..\.nuget\packages\azure.core\1.32.0\lib\net6.0\Azure.Core.dll</HintPath>
+    </Reference>
+    <Reference Include="Azure.Messaging.ServiceBus">
+      <HintPath>..\..\..\..\.nuget\packages\azure.messaging.servicebus\7.14.0\lib\netstandard2.0\Azure.Messaging.ServiceBus.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+
 </Project>
